refactor: Move private fn fetch_opts() into mod bare_index
To make it easier to turn git2 into an optional dep.
